+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.44.0
 
 * Add the "Unmanaged" KafkaTopic status update.
+* The `ContinueReconciliationOnManualRollingUpdateFailure` feature gate moves to beta stage and is enabled by default.
+  If needed, `ContinueReconciliationOnManualRollingUpdateFailure` can be disabled in the feature gates configuration in the Cluster Operator.
 
 ### Changes, deprecations and removals